refactor: add OptionController
